+++ MapboxCarPlayTests/CarPlayManagerTests.swift
@@ -120,6 +120,22 @@ class CarPlayManagerTests: XCTestCase {
         XCTAssertEqual(2, mapTemplate.leadingNavigationBarButtons.count)
         XCTAssertEqual(2, mapTemplate.trailingNavigationBarButtons.count)
     }
+    
+    func testManagerAsksDelegateForLeadingAndTrailingBarButtonsIfNotAvailable() {
+        // NOTE: Xcode is going to complain here - ignore. This is a known XCTest bug.
+        guard #available(iOS 12, *) else { return }
+        
+        simulateCarPlayConnection(manager!)
+        
+        guard let fakeInterfaceController = manager?.interfaceController else {
+            XCTFail("Dependencies not met! Bailing...")
+            return
+        }
+        
+        let mapTemplate: CPMapTemplate = fakeInterfaceController.rootTemplate as! CPMapTemplate
+        XCTAssertEqual(0, mapTemplate.leadingNavigationBarButtons.count)
+        XCTAssertEqual(0, mapTemplate.trailingNavigationBarButtons.count)
+    }
 
     func testManagerAsksDelegateForMapButtonsIfAvailable() {
         // NOTE: Xcode is going to complain here - ignore. This is a known XCTest bug.
@@ -139,6 +155,21 @@ class CarPlayManagerTests: XCTestCase {
 
         let mapTemplate: CPMapTemplate = fakeInterfaceController.rootTemplate as! CPMapTemplate
         XCTAssertEqual(1, mapTemplate.mapButtons.count)
+    }
+    
+    func testManagerAsksDelegateForMapButtonsIfNotAvailable() {
+        // NOTE: Xcode is going to complain here - ignore. This is a known XCTest bug.
+        guard #available(iOS 12, *) else { return }
+        
+        simulateCarPlayConnection(manager!)
+        
+        guard let fakeInterfaceController = manager?.interfaceController else {
+            XCTFail("Dependencies not met! Bailing...")
+            return
+        }
+        
+        let mapTemplate: CPMapTemplate = fakeInterfaceController.rootTemplate as! CPMapTemplate
+        XCTAssertEqual(4, mapTemplate.mapButtons.count)
     }
 
     func testManagerTellsDelegateWhenNavigationStartsAndEndsDueToArrival() {
@@ -222,16 +253,7 @@ class CarPlayManagerTests: XCTestCase {
     }
 }
 
-
-
-
-@available(iOS 12.0, *)
-func simulateCarPlayConnection(_ manager: CarPlayManager) {
-    let fakeInterfaceController = FakeCPInterfaceController(#function)
-    let fakeWindow = CPWindow()
-
-    manager.application(UIApplication.shared, didConnectCarInterfaceController: fakeInterfaceController, to: fakeWindow)
-}
+//MARK: -
 
 import Quick
 import Nimble
@@ -245,15 +267,92 @@ class CarPlayManagerSpec: QuickSpec {
         var delegate: TestCarPlayManagerDelegate?
 
         beforeEach {
-            manager = CarPlayManager()
+            let directionsSpy = DirectionsSpy(accessToken: "asdf")
+            manager = CarPlayManager(styles: nil, directions: directionsSpy, eventsManager: nil)
             delegate = TestCarPlayManagerDelegate()
             manager!.delegate = delegate
 
             simulateCarPlayConnection(manager!)
         }
 
-        describe("Starting a trip", {
+        //MARK: Previewing Routes
+        describe("Previewing routes", {
+            beforeEach {
+                manager!.mapTemplateProvider = MapTemplateSpyProvider()
+            }
 
+            let previewRoutesAction = {
+                let route = Fixture.route(from: "route-with-banner-instructions")
+                let waypoints = route.routeOptions.waypoints
+
+                let directionsSpy = manager!.directions as! DirectionsSpy
+
+                manager!.previewRoutes(for: route.routeOptions, completionHandler: {})
+                directionsSpy.fireLastCalculateCompletion(with: waypoints, routes: [route], error: nil)
+            }
+
+            context("when the trip is not customized by the developer", {
+                beforeEach {
+                    previewRoutesAction()
+                }
+                
+                it("previews a route/options with the default configuration") {
+                    let interfaceController = manager!.interfaceController as! FakeCPInterfaceController
+                    let mapTemplateSpy: MapTemplateSpy =  interfaceController.topTemplate as! MapTemplateSpy
+                    
+                    expect(mapTemplateSpy.currentTripPreviews).toNot(beEmpty())
+                    let expectedStartButtonTitle = NSLocalizedString("CARPLAY_GO", bundle: .mapboxNavigation, value: "Go", comment: "Title for start button in CPTripPreviewTextConfiguration")
+                    expect(mapTemplateSpy.currentPreviewTextConfiguration?.startButtonTitle).to(equal(expectedStartButtonTitle))
+                }
+            })
+            
+            context("when the delegate provides a custom trip", {
+                var customTrip: CPTrip!
+
+                beforeEach {
+                    let customTripDelegate = CustomTripPreviewDelegate()
+                    customTrip = CPTrip(origin: MKMapItem(), destination: MKMapItem(), routeChoices: [])
+                    customTripDelegate.customTrip = customTrip
+                    manager!.delegate = customTripDelegate
+
+                    previewRoutesAction()
+                }
+
+                it("shows trip previews for the custom trip") {
+                    let interfaceController = manager!.interfaceController as! FakeCPInterfaceController
+                    let mapTemplateSpy: MapTemplateSpy =  interfaceController.topTemplate as! MapTemplateSpy
+
+                    expect(mapTemplateSpy.currentTripPreviews).to(contain(customTrip))
+                    expect(mapTemplateSpy.currentPreviewTextConfiguration).toNot(beNil())
+                }
+            })
+            
+            context("when the delegate provides a custom trip preview text", {
+                var customTripPreviewTextConfiguration: CPTripPreviewTextConfiguration!
+                let customStartButtonTitleText = "Let's roll"
+                
+                beforeEach {
+                    let customTripDelegate = CustomTripPreviewDelegate()
+                    customTripPreviewTextConfiguration = CPTripPreviewTextConfiguration(startButtonTitle: customStartButtonTitleText, additionalRoutesButtonTitle: nil, overviewButtonTitle: nil)
+                    customTripDelegate.customTripPreviewTextConfiguration = customTripPreviewTextConfiguration
+                    manager!.delegate = customTripDelegate
+
+                    previewRoutesAction()
+                }
+                
+                it("previews a route/options with the custom trip configuration") {
+                    let interfaceController = manager!.interfaceController as! FakeCPInterfaceController
+                    let mapTemplateSpy: MapTemplateSpy =  interfaceController.topTemplate as! MapTemplateSpy
+                    
+                    expect(mapTemplateSpy.currentTripPreviews).toNot(beEmpty())
+                    expect(mapTemplateSpy.currentPreviewTextConfiguration?.startButtonTitle).to(equal(customStartButtonTitleText))
+                }
+
+            })
+        })
+        
+        //MARK: Starting a Trip
+        describe("Starting a trip", {
             let action = {
                 let fakeTemplate = CPMapTemplate()
                 let fakeRouteChoice = CPRouteChoice(summaryVariants: ["summary1"], additionalInformationVariants: ["addl1"], selectionSummaryVariants: ["selection1"])
@@ -297,6 +396,38 @@ class CarPlayManagerSpec: QuickSpec {
             })
         })
     }
+
+    private class CustomTripPreviewDelegate: CarPlayManagerDelegate {
+        var customTripPreviewTextConfiguration: CPTripPreviewTextConfiguration?
+        var customTrip: CPTrip?
+
+        func carPlayManager(_ carPlayManager: CarPlayManager, willPreview trip: CPTrip) -> (CPTrip) {
+            return customTrip ?? trip
+        }
+
+        func carPlayManager(_ carPlayManager: CarPlayManager, willPreview trip: CPTrip, with previewTextConfiguration: CPTripPreviewTextConfiguration) -> (CPTripPreviewTextConfiguration) {
+            return customTripPreviewTextConfiguration ?? previewTextConfiguration
+        }
+
+        func carPlayManager(_ carPlayManager: CarPlayManager, didBeginNavigationWith service: NavigationService) {
+            //no-op
+        }
+
+        func carPlayManagerDidEndNavigation(_ carPlayManager: CarPlayManager) {
+            //no-op
+        }
+    }
+}
+
+//MARK: -
+//MARK: Test Helper Methods
+
+@available(iOS 12.0, *)
+func simulateCarPlayConnection(_ manager: CarPlayManager) {
+    let fakeInterfaceController = FakeCPInterfaceController(#function)
+    let fakeWindow = CPWindow()
+    
+    manager.application(UIApplication.shared, didConnectCarInterfaceController: fakeInterfaceController, to: fakeWindow)
 }
 
 //MARK: Test Objects / Classes.
@@ -331,11 +462,11 @@ class TestCarPlayManagerDelegate: CarPlayManagerDelegate {
         return service
     }
 
-    func carPlayManager(_ carPlayManager: CarPlayManager, leadingNavigationBarButtonsCompatibleWith traitCollection: UITraitCollection, in template: CPTemplate, for activity: CarPlayActivity) -> [CPBarButton]? {
+    func carPlayManager(_ carPlayManager: CarPlayManager, leadingNavigationBarButtonsCompatibleWith traitCollection: UITraitCollection, in: CPTemplate, for activity: CarPlayActivity) -> [CPBarButton]? {
         return leadingBarButtons
     }
 
-    func carPlayManager(_ carPlayManager: CarPlayManager, trailingNavigationBarButtonsCompatibleWith traitCollection: UITraitCollection, in template: CPTemplate, for activity: CarPlayActivity) -> [CPBarButton]? {
+    func carPlayManager(_ carPlayManager: CarPlayManager, trailingNavigationBarButtonsCompatibleWith traitCollection: UITraitCollection, in: CPTemplate, for activity: CarPlayActivity) -> [CPBarButton]? {
         return trailingBarButtons
     }
 
@@ -380,6 +511,29 @@ class TestCarPlaySearchControllerDelegate: CarPlaySearchControllerDelegate {
     
     func popTemplate(animated: Bool) {
         interfaceController?.popTemplate(animated: animated)
+    }
+}
+
+@available(iOS 12.0, *)
+class MapTemplateSpy: CPMapTemplate {
+    private(set) var currentTripPreviews: [CPTrip]?
+    private(set) var currentPreviewTextConfiguration: CPTripPreviewTextConfiguration?
+
+    override func showTripPreviews(_ tripPreviews: [CPTrip], textConfiguration: CPTripPreviewTextConfiguration?) {
+        currentTripPreviews = tripPreviews
+        currentPreviewTextConfiguration = textConfiguration
+    }
+    
+    override func hideTripPreviews() {
+        currentTripPreviews = nil
+        currentPreviewTextConfiguration = nil
+    }
+}
+
+@available(iOS 12.0, *)
+public class MapTemplateSpyProvider: MapTemplateProvider {
+    override public func createMapTemplate() -> CPMapTemplate {
+        return MapTemplateSpy()
     }
 }