+++ MapboxCoreNavigation/EventDetails.swift
@@ -5,8 +5,54 @@ import UIKit
 import AVFoundation
 import MapboxDirections
 
+protocol EventDetails: Encodable {
+    var event: String? { get }
+    var created: Date { get }
+    var sessionIdentifier: String { get }
+}
 
-struct EventDetails: Encodable {
+struct PerformanceEventDetails: EventDetails {
+    let event: String?
+    let created: Date
+    let sessionIdentifier: String
+    var counters: [Counter] = []
+    var attributes: [Attribute] = []
+    
+    private enum CodingKeys: String, CodingKey {
+        case event
+        case created
+        case sessionIdentifier = "sessionId"
+        case counters
+        case attributes
+    }
+    
+    struct Counter: Encodable {
+        let name: String
+        let value: Double
+    }
+    
+    struct Attribute: Encodable {
+        let name: String
+        let value: String
+    }
+    
+    init(event: String?, session: SessionState, createdOn created: Date?) {
+        self.event = event
+        sessionIdentifier = session.identifier.uuidString
+        self.created = created ?? Date()
+    }
+    
+    func encode(to encoder: Encoder) throws {
+        var container = encoder.container(keyedBy: CodingKeys.self)
+        try container.encodeIfPresent(event, forKey: .event)
+        try container.encode(created.ISO8601, forKey: .created)
+        try container.encode(sessionIdentifier, forKey: .sessionIdentifier)
+        try container.encode(counters, forKey: .counters)
+        try container.encode(attributes, forKey: .attributes)
+    }
+}
+
+struct NavigationEventDetails: EventDetails {
     
     let audioType: String = AVAudioSession.sharedInstance().audioType
     let applicationState: UIApplicationState = UIApplication.shared.applicationState
@@ -286,12 +332,11 @@ extension RouteLegProgress: Encodable {
     }
 }
 
+enum EventDetailsError: Error {
+    case EncodingError(String)
+}
+
 extension EventDetails {
-    
-    enum EventDetailsError: Error {
-        case EncodingError(String)
-    }
-    
     func asDictionary() throws -> [String: Any] {
         let data = try JSONEncoder().encode(self)
         if let dictionary = try JSONSerialization.jsonObject(with: data, options: .allowFragments) as? [String: Any] {