+++ MapboxCoreNavigation/NavigationEventsManager.swift
@@ -2,6 +2,8 @@ import Foundation
 import MapboxMobileEvents
 import MapboxDirections
 
+let NavigationEventTypeRouteRetrieval = "mobile.performance_trace"
+
  /**
  The `EventsManagerDataSource` protocol declares values required for recording route following events.
  */
@@ -89,11 +91,11 @@ open class NavigationEventsManager: NSObject {
         mobileEventsManager.sendTurnstileEvent()
     }
     
-    func navigationCancelEvent(rating potentialRating: Int? = nil, comment: String? = nil) -> EventDetails? {
+    func navigationCancelEvent(rating potentialRating: Int? = nil, comment: String? = nil) -> NavigationEventDetails? {
         guard let dataSource = dataSource, let sessionState = sessionState else { return nil }
         
         let rating = potentialRating ?? MMEEventsManager.unrated
-        var event = EventDetails(dataSource: dataSource, session: sessionState, defaultInterface: usesDefaultUserInterface)
+        var event = NavigationEventDetails(dataSource: dataSource, session: sessionState, defaultInterface: usesDefaultUserInterface)
         event.event = MMEEventTypeNavigationCancel
         event.arrivalTimestamp = sessionState.arrivalTimestamp
         
@@ -107,18 +109,34 @@ open class NavigationEventsManager: NSObject {
         return event
     }
     
-    func navigationDepartEvent() -> EventDetails? {
+    func navigationRouteRetrievalEvent() -> PerformanceEventDetails? {
+        guard let sessionState = sessionState,
+            let responseEndDate = sessionState.currentRoute.responseEndDate,
+            let fetchStartDate = sessionState.currentRoute.fetchStartDate else {
+            return nil
+        }
+
+        var event = PerformanceEventDetails(event: NavigationEventTypeRouteRetrieval, session: sessionState, createdOn: sessionState.currentRoute.responseEndDate)
+        event.counters.append(PerformanceEventDetails.Counter(name: "elapsed_time",
+                                                              value: responseEndDate.timeIntervalSince(fetchStartDate)))
+        if let routeIdentifier = sessionState.currentRoute.routeIdentifier {
+            event.attributes.append(PerformanceEventDetails.Attribute(name: "route_uuid", value: routeIdentifier))
+        }
+        return event
+    }
+    
+    func navigationDepartEvent() -> NavigationEventDetails? {
         guard let dataSource = dataSource, let sessionState = sessionState else { return nil }
 
-        var event = EventDetails(dataSource: dataSource, session: sessionState, defaultInterface: usesDefaultUserInterface)
+        var event = NavigationEventDetails(dataSource: dataSource, session: sessionState, defaultInterface: usesDefaultUserInterface)
         event.event = MMEEventTypeNavigationDepart
         return event
     }
     
-    func navigationArriveEvent() -> EventDetails? {
+    func navigationArriveEvent() -> NavigationEventDetails? {
         guard let dataSource = dataSource, let sessionState = sessionState else { return nil }
 
-        var event = EventDetails(dataSource: dataSource, session: sessionState, defaultInterface: usesDefaultUserInterface)
+        var event = NavigationEventDetails(dataSource: dataSource, session: sessionState, defaultInterface: usesDefaultUserInterface)
         event.event = MMEEventTypeNavigationArrive
         return event
     }
@@ -131,10 +149,10 @@ open class NavigationEventsManager: NSObject {
         return eventDictionary
     }
     
-    func navigationFeedbackEvent(type: FeedbackType, description: String?) -> EventDetails? {
+    func navigationFeedbackEvent(type: FeedbackType, description: String?) -> NavigationEventDetails? {
         guard let dataSource = dataSource, let sessionState = sessionState else { return nil }
 
-        var event = EventDetails(dataSource: dataSource, session: sessionState, defaultInterface: usesDefaultUserInterface)
+        var event = NavigationEventDetails(dataSource: dataSource, session: sessionState, defaultInterface: usesDefaultUserInterface)
         event.event = MMEEventTypeNavigationFeedback
         
         event.userId = UIDevice.current.identifierForVendor?.uuidString
@@ -146,11 +164,11 @@ open class NavigationEventsManager: NSObject {
         return event
     }
     
-    func navigationRerouteEvent(eventType: String = MMEEventTypeNavigationReroute) -> EventDetails? {
+    func navigationRerouteEvent(eventType: String = MMEEventTypeNavigationReroute) -> NavigationEventDetails? {
         guard let dataSource = dataSource, let sessionState = sessionState else { return nil }
 
         let timestamp = Date()
-        var event = EventDetails(dataSource: dataSource, session: sessionState, defaultInterface: usesDefaultUserInterface)
+        var event = NavigationEventDetails(dataSource: dataSource, session: sessionState, defaultInterface: usesDefaultUserInterface)
         event.event = eventType
         if let lastRerouteDate = sessionState.lastRerouteDate {
             event.secondsSinceLastReroute = round(timestamp.timeIntervalSince(lastRerouteDate))
@@ -176,6 +194,12 @@ open class NavigationEventsManager: NSObject {
     public func sendCarPlayDisconnectEvent() {
         let date = Date()
         mobileEventsManager.enqueueEvent(withName: MMEventTypeNavigationCarplayDisconnect, attributes: [MMEEventKeyEvent: MMEventTypeNavigationCarplayDisconnect, MMEEventKeyCreated: date.ISO8601])
+        mobileEventsManager.flush()
+    }
+    
+    func sendRouteRetrievalEvent() {
+        guard let attributes = try? navigationRouteRetrievalEvent()?.asDictionary() else { return }
+        mobileEventsManager.enqueueEvent(withName: NavigationEventTypeRouteRetrieval, attributes: attributes ?? [:])
         mobileEventsManager.flush()
     }